Cleaned up code and moved content, alerts, and homepage into separate files.
